feat: 강사 마이페이지 화면 구현

+++ src/main/java/com/offcourse/mypage/model/dto/TeacherMyPageResponse.java
@@ -26,5 +26,5 @@ public class TeacherMyPageResponse {
     private Double averageRating;
     private Integer courseSize;
     private Integer courseCurrentSize;
-    private Double accountPrice;
+    private Integer accountPrice;
 }
